# fe-angular/src/app/core/template/openapi.yaml
openapi: 3.0.1
info:
  title: DSS API Docs V1
  description: DSS API Docs V1
  version: 1.0.0
security:
- bearerAuth: []
...
paths:
  /categories:
    get:
      tags:
      - Category
      summary: Get categories
      parameters:
      - name: pagenumber
        in: query
        description: The page number to retrieve
        schema:
          type: integer
          default: 1
      - name: pagesize
        in: query
        description: The number of items per page
        schema:
          type: integer
          default: 6
      responses:
        "200":
          description: The categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
                example:
                  success: true
                  payload:
                    data: []
        "500":
          description: When got server exception
          content:
            application/json:
              schema:
                type: string
                example: Internal server error
    post:
      tags:
      - Category
      summary: Create new category
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                score:
                  type: integer
                type:
                  type: string
              example:
                name: Tham gia tạo Proposal dự án >2M
                description: Tham gia tạo Proposal dự án >2M
                score: 200
                type: bizdev
        required: false
      responses:
        "200":
          description: The category created
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
        "422":
          description: "Unprocessable Entity, the data is not valid"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidatorErrorItem'
                example:
                  success: false
                  errors:
                  - value: ""
                    msg: name is invalid
                    param: name
                    location: body
        "500":
          description: When got server exception
          content:
            application/json:
              schema:
                type: string
                example: Internal server error
      x-codegen-request-body-name: body
...
  /submit-request:
    put:
      tags:
      - submitRequest
      summary: Update submit request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                accountName:
                  type: string
                descriptions:
                  type: string
                supervisor:
                  type: integer
                category:
                  type: integer
                hasNotifyMail:
                  type: boolean
              example:
                id: 1
                accountName: Join Smith
                descriptions: example@email.com
                supervisor: 1
                category: 1
                hasNotifyMail: true
        required: true
      responses:
        "200":
          description: The submit Request successfully
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  msg: Update success
        "500":
          description: When got server exception
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
      x-codegen-request-body-name: body
    post:
      tags:
      - submitRequest
      summary: Create submit request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accountName:
                  type: string
                descriptions:
                  type: string
                approver:
                  type: string
                supervisor:
                  type: integer
                category:
                  type: integer
                hasNotifyMail:
                  type: boolean
              example:
                accountName: Join Smith
                descriptions: example@email.com
                approver: tienpm3
                supervisor: 1
                category: 1
                hasNotifyMail: true
        required: true
      responses:
        "200":
          description: The submit Request successfully
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  payload:
                    suppervisorApproved: false
                    approverApproved: false
                    status: WAITING
                    id: 3
                    accountName: tienpm3
                    descriptions: params.descriptions
                    supervisor: 1
                    categoryId: 1
                    hasNotifyMail: true
                    createBy: 1
        "500":
          description: When got server exception
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
      x-codegen-request-body-name: body
...
  /users:
    get:
      tags:
      - User
      summary: Admin get all users
      parameters:
      - name: page
        in: query
        description: The page want to load
        schema:
          type: number
      - name: rowPerPage
        in: query
        description: The rowPerPage want to load
        schema:
          type: number
      - name: textSearch
        in: query
        description: The textSearch(by email) want to load
        schema:
          type: string
      responses:
        "200":
          description: The users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                example:
                  success: true
                  payload:
                    data: []
                    currentPage: 1
                    totalPage: 1
                    totalItems: 12
        "422":
          description: "Unprocessable Entity, the data is not valid"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidatorErrorItem'
                example:
                  success: false
                  errors:
                  - msg: Page number must be a positive integer
                    param: page
                  - msg: Row per page must be a positive integer not larger than 200
                    param: rowPerPage
        "500":
          description: When got server exception
          content:
            application/json:
              schema:
                type: string
                example: Internal server error
    put:
      tags:
      - User
      summary: Update user profile
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                identityCardFront:
                  type: string
                  description: "The user identity card front face, upload type jpg|jpeg|png"
                  format: binary
                identityCardBack:
                  type: string
                  description: "The user identity card back face, upload type jpg|jpeg|png"
                  format: binary
                avatar:
                  type: string
                  description: "The user avatar, upload type jpg|jpeg|png"
                  format: binary
                cover:
                  type: string
                  description: "The user cover, upload type jpg|jpeg|png"
                  format: binary
                data:
                  type: string
                  description: The body stringify information
                  default: "{\"fullName\": \"David Hall\", \"email\": \"email@domain.com\"\
                    , \"phone\": \"054888383223\"}"
      responses:
        "200":
          description: The user profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "304":
          description: Not Modified
          content: {}
        "403":
          description: When data cannot be process
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidatorErrorItem'
                example:
                  success: false
                  errors:
                  - msg: "User avatar file is invalid, only image files are allowed!"
                    param: userAvatarInvalid
                    location: body
                  - msg: Email is not available
                    param: emailNotAvailable
                    location: body
        "422":
          description: "Unprocessable Entity, the data is not valid"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidatorErrorItem'
                example:
                  success: false
                  errors:
                  - value: mail@mail
                    msg: User email is invalid
                    param: email
                    location: body
        "500":
          description: When got server exception
          content: {}
...
components:
  schemas:
    ValidatorErrorItem:
      type: object
      properties:
        value:
          type: string
          description: The value got on request
        msg:
          type: string
          description: The error message
        param:
          type: string
          description: The key of value
        location:
          type: string
          description: The location of value
...
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the user
        firstName:
          type: string
          description: first Name of the user
        lastName:
          type: string
          description: last Name of the user
        email:
          type: string
          description: Email address of the user
    Category:
      type: object
      properties:
        name:
          type: string
          description: The name of the category
        description:
          type: string
          description: A short description of the category
        score:
          type: number
          description: The score associated with the category
          format: float
        type:
          type: string
          description: The type of the category
  securitySchemes:
    bearerAuth:
      type: apiKey
      description: "Enter your Bearer token in the format: Bearer <token>"
      name: Authorization
      in: header